search: concurrent evaluation of OR queries
This commit makes our evaluation of OR queries concurrent, which should
achieve a significant speed-up.

As documented in the change, there are two existing bugs that are not
fixed by this change, notably:

1. Stats aren't updated to reflect the restricted set of repos when we stop searching due to hitting the desired result count.
2. Alerts generated by each evaluated term are ignored.

Doing deduplication and stats adjustment at this layer seems
sub-optimal. I'd like to discuss with @rvantonder the idea of making
this query evaluation layer work with our `streaming.Sender` interface,
and leveraging `run.Aggregator` for these concerns.